Add files via upload
We can send the email to all of the listed leads by preparing or update data in the CSV file only. Kindly change the date inside the code to be the same as the date on CSV file, it will help us in sending email to the leads based on the targeted date.